API reference

Smesh.build_delaunay_triangulation — Method

Smesh.build_polygon_mesh — Method

build_polygon_mesh(data_points, triangulation_vertices; mesh_type=:standard_voronoi, orthogonal_boundary_edges=true)

There are three different mesh types:

:standard_voronoi => standard voronoi, but use centroid if the circumcenter lies outside the triangle

:centroids => not an actual voronoi, always use centroids and not circumcenters as vertices for the mesh

:incenters => not an actual voronoi, always use incenters and not circumcenters as vertices for the mesh

:pure_voronoi => pur Voronoi mesh (just for experiments, should not be used for computation)

Smesh.delaunay_compute_neighbors — Method

Smesh.voronoi_compute_neighbors — Method